+++ OrbitCore/EventBuffer.h
@@ -16,7 +16,7 @@
 
 //-----------------------------------------------------------------------------
 struct CallstackEvent {
-  CallstackEvent(long long a_Time = 0, CallstackID a_Id = 0, ThreadID a_TID = 0)
+  CallstackEvent(uint64_t a_Time = 0, CallstackID a_Id = 0, ThreadID a_TID = 0)
       : m_Time(a_Time), m_Id(a_Id), m_TID(a_TID) {}
   long long m_Time;
   CallstackID m_Id;
@@ -37,12 +37,12 @@ class EventBuffer {
     m_MinTime = LLONG_MAX;
     m_MaxTime = 0;
   }
-  std::map<ThreadID, std::map<long long, CallstackEvent> >& GetCallstacks() {
+  std::map<ThreadID, std::map<uint64_t, CallstackEvent> >& GetCallstacks() {
     return m_CallstackEvents;
   }
   Mutex& GetMutex() { return m_Mutex; }
-  std::vector<CallstackEvent> GetCallstackEvents(long long a_TimeBegin,
-                                                 long long a_TimeEnd,
+  std::vector<CallstackEvent> GetCallstackEvents(uint64_t a_TimeBegin,
+                                                 uint64_t a_TimeEnd,
                                                  ThreadID a_ThreadId = 0);
   long long GetMaxTime() const { return m_MaxTime; }
   long long GetMinTime() const { return m_MinTime; }
@@ -60,16 +60,16 @@ class EventBuffer {
 #endif
 
   //-----------------------------------------------------------------------------
-  void RegisterTime(long long a_Time) {
+  void RegisterTime(uint64_t a_Time) {
     if (a_Time > m_MaxTime) m_MaxTime = a_Time;
     if (a_Time > 0 && a_Time < m_MinTime) m_MinTime = a_Time;
   }
 
   //-----------------------------------------------------------------------------
-  void AddCallstackEvent(long long a_Time, CallstackID a_CSHash,
+  void AddCallstackEvent(uint64_t a_Time, CallstackID a_CSHash,
                          ThreadID a_TID) {
     ScopeLock lock(m_Mutex);
-    std::map<long long, CallstackEvent>& threadMap = m_CallstackEvents[a_TID];
+    std::map<uint64_t, CallstackEvent>& threadMap = m_CallstackEvents[a_TID];
     threadMap[a_Time] = CallstackEvent(a_Time, a_CSHash, a_TID);
     RegisterTime(a_Time);
   }
@@ -78,7 +78,7 @@ class EventBuffer {
 
  private:
   Mutex m_Mutex;
-  std::map<ThreadID, std::map<long long, CallstackEvent> > m_CallstackEvents;
-  std::atomic<long long> m_MaxTime;
-  std::atomic<long long> m_MinTime;
+  std::map<ThreadID, std::map<uint64_t, CallstackEvent> > m_CallstackEvents;
+  std::atomic<uint64_t> m_MaxTime;
+  std::atomic<uint64_t> m_MinTime;
 };
